separare CI/demo/tests. Shaping up for PR
