+++ .gitignore
@@ -1,2 +1,4 @@
 *.pyc
 *.swp
+build
+tests/.cache
